requirements.txt
sentence-transformers
faiss-cpu
rapidfuzz
python-dotenv
paddleocr 
opencv-python 
pillow
paddlepaddle
streamlit
